- Mac OS X support - Docker devcontainer - Bugfixes
