+++ .gitignore
@@ -1,4 +1,4 @@
 bazel-*
 .bazelrc.user
-.ijwb/
+/.ijwb/
 /.idea/
