+++ examples/test/.bazelrc
@@ -1,2 +1,3 @@
 #build --disk_cache=/tmp/bazel_disk_cache
 build --repository_cache=/tmp/bazel_repo_cache
+build --java_runtime_version=remotejdk_11
